Cleanup
Stripped example mod down to a single demonstration of use. Moved it to src>test-java>(default package).

Hopefully fixed whitespace issues.

Ran gradlew licenseFormat, which for some reason affected EntityTeleportEvent and a few other random files?
