+++ .gitignore
@@ -44,4 +44,3 @@ changelog.txt
 # Srg2Source causes import bugs on these patches. If you intentionnally change these, remove the ignore.
 /patches/minecraft/net/minecraft/client/renderer/ViewFrustum.java.patch
 /patches/minecraft/net/minecraft/data/BlockModelDefinition.java.patch
-*.xml
