+++ src/main/resources/META-INF/mods.toml
@@ -15,4 +15,3 @@ license="LGPL v2.1"
     Forge, a broad compatibility API.
     '''
 		
-
